refactor: コンポーネントのディレクトリ変更
適切なディレクトリに配置するため
